+++ src/test/java/com/epam/reportportal/service/statistics/StatisticsRunnable.java
@@ -23,7 +23,8 @@ import static org.hamcrest.Matchers.instanceOf;
 
 public class StatisticsRunnable {
 	public static void main(String... args) throws ClassNotFoundException {
-		StatisticsService service = new StatisticsService(TestUtils.standardParameters());
-		assertThat(service.getStatistics(), instanceOf(Class.forName(args[0])));
+		try(StatisticsService service = new StatisticsService(TestUtils.standardParameters())) {
+			assertThat(service.getStatistics(), instanceOf(Class.forName(args[0])));
+		}
 	}
 }